+++ src/datasource/CustomerDatasource.ts
@@ -45,6 +45,25 @@ const getCusCorporate = async (params: object) => {
     })
     .catch((err) => console.log(err));
 };
+const checkTaxNo = async (taxNo: string) => {
+  return await httpClient
+    .get(`${baseUrl}/customer/get-customer-tax/${taxNo}`)
+    .then((res: AxiosResponse) => res.data)
+    .catch((err) => console.log(err));
+};
+
+const createCustomerEx = async (params: object) => {
+  return await httpClient
+    .post(`${baseUrl}/customer/create-customer-ex`, params)
+    .then((res: AxiosResponse) => res.data)
+    .catch((err) => console.log(err));
+};
+const updateCustomerEx = async (data: object) => {
+  return await httpClient
+    .patch(`${baseUrl}/customer/update-customer-ex`, data)
+    .then((res: AxiosResponse) => res.data)
+    .catch((err) => console.log(err));
+};
 
 export {
   getCustomers,
@@ -53,4 +72,7 @@ export {
   checkPhoneByShop,
   checkPhoneAllShop,
   getCusCorporate,
+  checkTaxNo,
+  createCustomerEx,
+  updateCustomerEx,
 };