+++ src/pages/ShopManagementPage/CorporateShop/CreateCorporateShop.tsx
@@ -18,6 +18,16 @@ import Button from "../../../components/Button/Button";
 import { GoogleMap, MarkerF, useLoadScript } from "@react-google-maps/api";
 import { getMasterAddress, getProvince } from "../../../datasource/AddressDatasource";
 import _ from "lodash";
+import {
+  createCustomerEx,
+  getCustomersById,
+  updateCustomerEx,
+} from "../../../datasource/CustomerDatasource";
+import { getProductBrand } from "../../../datasource/ProductDatasource";
+import { zoneDatasource } from "../../../datasource/ZoneDatasource";
+import DatePicker from "../../../components/DatePicker/DatePicker";
+import dayjs from "dayjs";
+import { useNavigate } from "react-router-dom";
 
 const UploadVeritical = styled(Upload)`
   .ant-upload,
@@ -66,12 +76,15 @@ export const AntSelectCustom = styled(Select)`
 `;
 
 export const CreateCorporateShop: React.FC = () => {
+  const navigate = useNavigate();
   const userProfile = JSON.parse(localStorage.getItem("profile")!);
+  const company = JSON.parse(localStorage.getItem("company")!);
   const [form1] = useForm();
   const [form2] = useForm();
   const { pathname } = window.location;
   const pathSplit = pathname.split("/") as Array<string>;
   const isEditing = pathSplit[3] !== "create";
+  const id = isEditing ? pathSplit[3] : pathSplit[4];
 
   const [current, setCurrent] = useState(0);
   const [imageUrl, setImageUrl] = useState<string>();
@@ -82,6 +95,9 @@ export const CreateCorporateShop: React.FC = () => {
   const [provinceList, setProvinceList] = useState<any>([]);
   const [districtList, setDistrictList] = useState<any>([]);
   const [subdistrictList, setSubdistrictList] = useState<any>([]);
+
+  const [brand, setBrand] = useState<any>([]);
+  const [zone, setZone] = useState<any>([]);
 
   const { isLoaded } = useLoadScript({
     googleMapsApiKey: "AIzaSyDg4BI3Opn-Bo2Pnr40Z7PKlC6MOv8T598",
@@ -101,6 +117,171 @@ export const CreateCorporateShop: React.FC = () => {
     },
     [center],
   );
+
+  const getCusComById = async () => {
+    if (isEditing) {
+      await getCustomersById(id).then(async (res) => {
+        await form1.setFieldsValue({
+          customerType: res?.customerType,
+          customerNo: res?.customerNo,
+          zone: res?.zone,
+          brand: res?.productBrand.map((b) => b.product_brand_name),
+          isActive: res?.isActive,
+        });
+        await form2.setFieldsValue({
+          customerId: res.customerId,
+          title: res?.customer.title,
+          ownerFirstname: res?.customer.ownerFirstname,
+          ownerLastname: res?.customer.ownerLastname,
+          telephone: res?.customer.telephone,
+          email: res?.customer.email,
+          nickname: res?.customer.nickname,
+          createDate: dayjs(res.createDate),
+          customerName: res?.customerName,
+          taxNo: res?.customer.taxNo,
+          termPayment: res?.termPayment === "COD" ? res.termPayment : "N",
+          creditDay: res?.termPayment === "COD" ? "" : res.termPayment.split("N")[1],
+          provinceId:
+            res?.customer.provinceId ||
+            (await getMasterProvince(res?.customer?.province).then((res) => {
+              return res.provinceId;
+            })),
+          provinceName: res?.customer.province,
+          districtId: res?.customer.districtId,
+          districtName: res?.customer.district,
+          subdistrictId: res?.customer.subdistrictId,
+          subdistrictName: res?.customer.subdistrict || "",
+          postcode: res?.customer.postcode,
+          address: res?.customer.address,
+          lat: res?.customer.lat,
+          lag: res?.customer.lag,
+          customerCompanyId: res?.customerCompanyId,
+          masterAddressId: res?.customer?.masterAddressId,
+          isActive: res.isActive,
+        });
+      });
+    } else {
+      id !== "0" &&
+        (await getCustomersById(id).then(async (res) => {
+          console.log("data", res);
+          let masterAddress: any = "";
+          let masterProvince: any = "";
+          if (res?.customer?.province) {
+            masterProvince = await getMasterProvince(res?.customer?.province).then((res) => res);
+            masterAddress = await getAddress(
+              masterProvince?.provinceId,
+              res?.customer?.district || "",
+              res?.customer?.subdistrict || "",
+            ).then((res) => res);
+          }
+          await form1.setFieldValue("isActive", true);
+          await form2.setFieldsValue({
+            customerId: res.customerId,
+            title: res.customer.title,
+            ownerFirstname: res.customer.ownerFirstname,
+            ownerLastname: res.customer.ownerLastname,
+            nickname: res.customer.nickname,
+            taxNo: res.customer.taxNo,
+            telephone: res.customer.telephone,
+            email: res.customer.email,
+            createDate: dayjs(res.createDate),
+            customerName: res.customerName,
+            address: res.customer.address,
+            lat: res.customer.lat,
+            lag: res.customer.lag,
+            provinceId: masterProvince.provinceId,
+            provinceName: masterProvince.provinceName,
+            districtName: masterAddress?.districtId?.districtName || "",
+            subdistrictName: masterAddress?.subdistrictId?.subdistrictName || "",
+            postcode: masterAddress?.subdistrictId?.postCode,
+            isActive: res.isActive,
+            masterAddress: masterAddress,
+          });
+        }));
+    }
+  };
+  const getMasterProvince = async (proNav?: string) => {
+    const province = await getProvince().then((res) => {
+      return res.responseData;
+    });
+    setProvinceList(province || []);
+    const provinceId = province.find((y: any) => {
+      const map = proNav?.endsWith(y.provinceName);
+      if (map) {
+        return { provinceId: y.provinceId, provinceName: y.provinceName };
+      }
+    });
+    return provinceId;
+  };
+  const getAddress = async (proId: number, dis: string, sub: string) => {
+    const getAddr = await getMasterAddress(proId).then((res) => {
+      return res.responseData;
+    });
+    setMasterAddr(getAddr);
+    const groupsDistrict = _.groupBy(getAddr || [], "districtName");
+    const objDistrict = _.map(groupsDistrict, (items, disName) => {
+      return {
+        districtId: items.find((item: any) => item.districtName === disName).districtId,
+        districtName: disName,
+      };
+    });
+    setDistrictList(objDistrict);
+    const districtId = await objDistrict.find((y: any) => {
+      const map = dis.endsWith(y.districtName);
+      if (map) {
+        return y.districtId;
+      }
+    });
+
+    const groupsSubdistrict = _.groupBy(
+      getAddr.filter((y: any) => dis.endsWith(y.districtName)) || [],
+      "subdistrictName",
+    );
+    const objSubdistrict = _.map(groupsSubdistrict, (items, disName) => {
+      return {
+        masterAddressId: items.find((item: any) => item.subdistrictName === disName)
+          .masterAddressId,
+        subdistrictId: items.find((item: any) => item.subdistrictName === disName).subdistrictId,
+        subdistrictName: disName,
+        postCode: items.find((item: any) => item.subdistrictName === disName).postcode,
+      };
+    });
+    setSubdistrictList(objSubdistrict);
+    const subdistrictId = await objSubdistrict.find((y: any) => {
+      const map = sub.endsWith(y.subdistrictName);
+      if (map) {
+        return y.subdistrictId;
+      }
+    });
+    return { districtId: districtId, subdistrictId: subdistrictId };
+  };
+  const getBrand = async () => {
+    await getProductBrand(company?.companyCode).then((res) => {
+      const mapBrand = res.map((b) => ({
+        company: b.company,
+        product_brand_id: b.productBrandId,
+        product_brand_logo: b.productBrandLogo,
+        product_brand_name: b.productBrandName,
+      }));
+      setBrand(mapBrand);
+    });
+  };
+  const getZone = async () => {
+    await zoneDatasource.getAllZoneByCompany(company?.companyCode).then((res) => {
+      setZone(res);
+    });
+  };
+  const onChangeMap = () => {
+    const payload = form2.getFieldsValue(true);
+    setCenter({ lat: Number(payload.lat), lng: Number(payload.lag) });
+  };
+
+  useEffect(() => {
+    getMasterProvince();
+    getBrand();
+    getZone();
+    getCusComById();
+  }, []);
 
   const staticData = [
     {
@@ -125,45 +306,6 @@ export const CreateCorporateShop: React.FC = () => {
     },
   ];
 
-  const getMasterProvince = async () => {
-    const province = await getProvince().then((res) => {
-      return res.responseData;
-    });
-    setProvinceList(province || []);
-  };
-  const onChangeDistrict = async () => {
-    const payload = form2.getFieldsValue(true);
-    const groupsSubdistrict = _.groupBy(
-      masterAddr.filter((y: any) => payload.districtId.endsWith(y.districtName)) || [],
-      "subdistrictName",
-    );
-    const objSubdistrict = _.map(groupsSubdistrict, (items, disName) => {
-      return {
-        masterAddressId: items.find((item: any) => item.subdistrictName === disName)
-          .masterAddressId,
-        subdistrictId: items.find((item: any) => item.subdistrictName === disName).subdistrictId,
-        subdistrictName: disName,
-        postCode: items.find((item: any) => item.subdistrictName === disName).postcode,
-      };
-    });
-    form2.setFieldsValue({ subdistrictId: "", postcode: "" });
-    setSubdistrictList(objSubdistrict);
-  };
-  const onChangeSubdistrict = async () => {
-    const payload = form2.getFieldsValue(true);
-    const getSubdistrict = masterAddr.find((x: any) =>
-      payload.subdistrictId.endsWith(x.subdistrictName),
-    );
-    form2.setFieldsValue({
-      masterAddressId: getSubdistrict.masterAddressId,
-      postcode: getSubdistrict ? getSubdistrict.postcode || "" : "",
-    });
-  };
-
-  useEffect(() => {
-    getMasterProvince();
-  }, []);
-
   const onChangeProvince = async () => {
     const payload = form2.getFieldsValue(true);
     const getAddr = await getMasterAddress(payload.provinceId).then((res) => {
@@ -177,8 +319,46 @@ export const CreateCorporateShop: React.FC = () => {
         districtName: disName,
       };
     });
-    form2.setFieldsValue({ districtId: "", subdistrictId: "", postcode: "" });
+    form2.setFieldsValue({
+      districtName: "",
+      subdistrictName: "",
+      postcode: "",
+    });
     setDistrictList(objDistrict);
+  };
+  const onChangeDistrict = async () => {
+    const payload = form2.getFieldsValue(true);
+    const groupsSubdistrict = _.groupBy(
+      masterAddr.filter((y: any) => payload.districtName.endsWith(y.districtName)) || [],
+      "subdistrictName",
+    );
+    const objSubdistrict = _.map(groupsSubdistrict, (items, disName) => {
+      return {
+        masterAddressId: items.find((item: any) => item.subdistrictName === disName)
+          .masterAddressId,
+        subdistrictId: items.find((item: any) => item.subdistrictName === disName).subdistrictId,
+        subdistrictName: disName,
+        postCode: items.find((item: any) => item.subdistrictName === disName).postcode,
+      };
+    });
+    form2.setFieldsValue({ subdistrictName: "", postcode: "" });
+    setSubdistrictList(objSubdistrict);
+  };
+  const onChangeSubdistrict = async () => {
+    const payload = form2.getFieldsValue(true);
+    const getSubdistrict = masterAddr.find((x: any) =>
+      payload.subdistrictName.endsWith(x.subdistrictName),
+    );
+    form2.setFieldsValue({
+      masterAddressId: getSubdistrict.masterAddressId,
+      provinceId: getSubdistrict.provinceId,
+      provinceName: getSubdistrict.provinceName,
+      districtId: getSubdistrict.districtId,
+      districtName: getSubdistrict.districtName,
+      subdistrictId: getSubdistrict.subdistrictId,
+      subdistrictName: getSubdistrict.subdistrictName,
+      postcode: getSubdistrict ? getSubdistrict.postcode || "" : "",
+    });
   };
 
   const PageTitle = () => {
@@ -219,14 +399,14 @@ export const CreateCorporateShop: React.FC = () => {
         return <StepOne />;
       }
       case 1: {
-        return <StepTwo />;
+        return StepTwo();
       }
     }
   };
   const StepOne = () => {
     return (
       <Form form={form1} layout='vertical'>
-        <Row justify={"space-between"} gutter={16}>
+        {/* <Row justify={"space-between"} gutter={16}>
           <Col span={24}>
             <Row>
               <Col span={3}>
@@ -299,11 +479,11 @@ export const CreateCorporateShop: React.FC = () => {
             </Row>
           </Col>
         </Row>
-        <br />
+        <br /> */}
         <Row>
           <Col span={10}>
             <Form.Item
-              name='type'
+              name='customerType'
               label='ประเภทคู่ค้า'
               rules={[
                 {
@@ -314,8 +494,8 @@ export const CreateCorporateShop: React.FC = () => {
             >
               <Radio.Group style={{ width: "100%" }}>
                 <Space direction='vertical'>
-                  <Radio value='Dealer'>Dealer</Radio>
-                  <Radio value='Sub-Dealer'>Sub-Dealer</Radio>
+                  <Radio value='DL'>Dealer</Radio>
+                  <Radio value='SD'>Sub-Dealer</Radio>
                 </Space>
               </Radio.Group>
             </Form.Item>
@@ -324,22 +504,65 @@ export const CreateCorporateShop: React.FC = () => {
         <br />
         <Row justify={"space-between"} gutter={16}>
           <Col span={12}>
-            <Form.Item name='shopNo' label='รหัสร้านค้า'>
+            <Form.Item name='customerNo' label='รหัสร้านค้า'>
               <Input placeholder='ระบุรหัสร้านค้า' autoComplete='off' />
             </Form.Item>
           </Col>
           <Col span={12}>
             <Form.Item name='zone' label='เขต'>
-              <Selects placeholder='เลือกเขต' data={[]} />
+              <Selects
+                allowClear
+                placeholder='เลือกเขต'
+                data={
+                  zone.map((z) => ({
+                    label: z.zoneName,
+                    key: z.zoneId,
+                    value: z.zoneName,
+                  })) || []
+                }
+              />
             </Form.Item>
           </Col>
         </Row>
         <Row justify={"space-between"} gutter={16}>
           <Col span={12}>
             <Form.Item name='brand' label='แบรนด์สินค้า (Product Brands)'>
-              <Selects placeholder='เลือก Product Brands' data={[]} />
+              <Selects
+                mode='multiple'
+                placeholder='เลือก Product Brands'
+                data={
+                  brand.map((z) => ({
+                    label: z.product_brand_name,
+                    key: z.product_brand_id,
+                    value: z.product_brand_name,
+                  })) || []
+                }
+              />
             </Form.Item>
           </Col>
+        </Row>
+        <Row>
+          {isEditing && (
+            <Col span={10}>
+              <Form.Item
+                name='isActive'
+                label='สถานะ'
+                rules={[
+                  {
+                    required: true,
+                    message: "*โปรดระบุสถานะ",
+                  },
+                ]}
+              >
+                <Radio.Group style={{ width: "100%" }}>
+                  <Space direction='vertical'>
+                    <Radio value={true}>เปิดใช้งาน</Radio>
+                    <Radio value={false}>ปิดใช้งาน</Radio>
+                  </Space>
+                </Radio.Group>
+              </Form.Item>
+            </Col>
+          )}
         </Row>
       </Form>
     );
@@ -375,7 +598,7 @@ export const CreateCorporateShop: React.FC = () => {
         </Col>
         <Col span={6}>
           <Form.Item
-            name='firstname'
+            name='ownerFirstname'
             label='ชื่อจริง'
             rules={[
               {
@@ -389,7 +612,7 @@ export const CreateCorporateShop: React.FC = () => {
         </Col>
         <Col span={6}>
           <Form.Item
-            name='lastname'
+            name='ownerLastname'
             label='นามสกุล'
             rules={[
               {
@@ -402,7 +625,7 @@ export const CreateCorporateShop: React.FC = () => {
           </Form.Item>
         </Col>
         <Col span={6}>
-          <Form.Item name='idCard' label='หมายเลขบัตรประชาชน'>
+          <Form.Item name='nickname' label='ชื่อเล่น'>
             <Input autoComplete='off' />
           </Form.Item>
         </Col>
@@ -415,16 +638,12 @@ export const CreateCorporateShop: React.FC = () => {
       >
         <Col span={12}>
           <Form.Item
-            name='telephoneFirst'
+            name='telephone'
             label='เบอร์โทรศัพท์'
             rules={[
               {
                 required: true,
                 message: "กรุณากรอกเบอร์โทร",
-              },
-              {
-                pattern: /^[0-9.]*$/,
-                message: "กรุณากรอกตัวเลขเท่านั้น",
               },
             ]}
           >
@@ -444,8 +663,8 @@ export const CreateCorporateShop: React.FC = () => {
         }}
       >
         <Col span={12}>
-          <Form.Item name='startDate' label='วันที่เริ่มเป็นสมาชิก'>
-            <Input autoComplete='off' />
+          <Form.Item name='createDate' label='วันที่เริ่มเป็นสมาชิก'>
+            <DatePicker style={{ width: "100%" }} enablePast />
           </Form.Item>
         </Col>
       </Row>
@@ -468,7 +687,7 @@ export const CreateCorporateShop: React.FC = () => {
       >
         <Col span={12}>
           <Form.Item
-            name='shopName'
+            name='customerName'
             label='ชื่อร้านค้า'
             rules={[
               {
@@ -511,10 +730,54 @@ export const CreateCorporateShop: React.FC = () => {
               },
             ]}
           >
-            <Radio.Group style={{ width: "100%" }}>
-              <Space direction='vertical'>
+            <Radio.Group
+              style={{ width: "100%" }}
+              onChange={() => {
+                form2.setFieldValue("creditDay", "");
+              }}
+            >
+              <Space direction='vertical' style={{ width: "100%" }}>
                 <Radio value='COD'>เงินสด</Radio>
-                <Radio value='N'>เครดิต</Radio>
+                <Row gutter={8} style={{ width: "100%" }}>
+                  <Col>
+                    <Radio value='N'>เครดิต</Radio>
+                  </Col>
+                  <Col>
+                    <Form.Item
+                      noStyle
+                      shouldUpdate={(prev, current) => {
+                        return prev.termPayment !== current.termPayment;
+                      }}
+                      rules={[
+                        {
+                          required: true,
+                          message: "*โปรดระบุจำนวนวัน",
+                        },
+                      ]}
+                    >
+                      {({ getFieldValue }) => {
+                        return (
+                          getFieldValue("termPayment") === "N" && (
+                            <Col>
+                              <Form.Item
+                                noStyle
+                                name='creditDay'
+                                rules={[
+                                  {
+                                    required: true,
+                                    message: "*โปรดระบุจำนวนวัน",
+                                  },
+                                ]}
+                              >
+                                <Input suffix='วัน' />
+                              </Form.Item>
+                            </Col>
+                          )
+                        );
+                      }}
+                    </Form.Item>
+                  </Col>
+                </Row>
               </Space>
             </Radio.Group>
           </Form.Item>
@@ -569,7 +832,7 @@ export const CreateCorporateShop: React.FC = () => {
         </Col>
         <Col span={6}>
           <Form.Item
-            name='districtId'
+            name='districtName'
             label='อำเภอ/เขต'
             rules={[
               {
@@ -603,7 +866,7 @@ export const CreateCorporateShop: React.FC = () => {
         </Col>
         <Col span={6}>
           <Form.Item
-            name='subdistrictId'
+            name='subdistrictName'
             label='ตำบล/แขวง'
             rules={[
               {
@@ -658,7 +921,7 @@ export const CreateCorporateShop: React.FC = () => {
       >
         <Col span={12}>
           <Form.Item
-            name='addressDesc'
+            name='address'
             label='ที่อยู่ (บ้านเลขที่ หมู่ ซอย ชั้น อาคาร)'
             rules={[
               {
@@ -679,7 +942,7 @@ export const CreateCorporateShop: React.FC = () => {
       >
         <Col span={12}>
           <Form.Item
-            name='latitude'
+            name='lat'
             label='ตำแหน่ง Latitude'
             rules={[
               {
@@ -692,12 +955,12 @@ export const CreateCorporateShop: React.FC = () => {
               },
             ]}
           >
-            <Input autoComplete='off' />
+            <Input autoComplete='off' onChange={onChangeMap} />
           </Form.Item>
         </Col>
         <Col span={12}>
           <Form.Item
-            name='longitude'
+            name='lag'
             label='ตำแหน่ง Longitude'
             rules={[
               {
@@ -710,7 +973,7 @@ export const CreateCorporateShop: React.FC = () => {
               },
             ]}
           >
-            <Input autoComplete='off' />
+            <Input autoComplete='off' onChange={onChangeMap} />
           </Form.Item>
         </Col>
       </Row>
@@ -733,7 +996,6 @@ export const CreateCorporateShop: React.FC = () => {
       </Form>
     );
   };
-
   const footer = () => {
     return (
       <Row justify='space-between' gutter={12}>
@@ -760,11 +1022,91 @@ export const CreateCorporateShop: React.FC = () => {
     );
   };
 
-  const nextStep = () => {
+  const nextStep = async () => {
     if (current === 0) {
-      setCurrent(current + 1);
+      await form1
+        .validateFields()
+        .then((f1) => {
+          setCurrent(current + 1);
+        })
+        .catch((errInfo) => {
+          console.log("form1 errInfo", errInfo);
+        });
     } else {
-      console.log(1);
+      form2
+        .validateFields()
+        .then(async (f) => {
+          const f1 = form1.getFieldsValue(true);
+          const f2 = form2.getFieldsValue(true);
+
+          const payload: any = {};
+          const cusCom: any = {};
+          const jsonBrand = f1.brand.map((x) => {
+            const find = brand.find((y) => y.product_brand_name === x);
+            if (find) {
+              return find;
+            }
+          });
+          cusCom.customerId = f2.customerId;
+          cusCom.isNav = id !== "0" ? true : false;
+          cusCom.customerName = f2.customerName;
+          cusCom.customerNo = f1.customerNo;
+          cusCom.company = company.companyCode;
+          cusCom.customerType = f1.customerType;
+          cusCom.zone = f1.zone;
+          cusCom.termPayment =
+            f2.termPayment === "N" ? f2.termPayment + f2.creditDay : f2.termPayment;
+          cusCom.creditLimit = 0;
+          cusCom.isActive = f1.isActive;
+          cusCom.salePersonCode = "";
+          cusCom.updateBy = userProfile.firstname + " " + userProfile.lastname;
+          cusCom.productBrand = jsonBrand;
+
+          payload.customerId = f2.customerId;
+          payload.address = f2.address;
+          payload.masterAddressId =
+            f2.masterAddress?.subdistrictId?.masterAddressId || f2.masterAddressId;
+          payload.provinceId = f2?.provinceId;
+          payload.districtId = f2?.masterAddress?.districtId?.districtId || f2?.districtId;
+          payload.subdistrictId =
+            f2?.masterAddress?.subdistrictId?.subdistrictId || f2?.subdistrictId;
+          payload.province = f2?.provinceName;
+          payload.district = f2?.districtName;
+          payload.subdistrict = f2?.subdistrictName;
+          payload.postcode = f2?.postcode;
+          payload.telephone = f2?.telephone;
+          payload.taxNo = f2.taxNo;
+          payload.updateBy = userProfile.firstname + " " + userProfile.lastname;
+          payload.lat = f2.lat;
+          payload.lag = f2.lag;
+          payload.statusMainTel = true;
+          payload.statusSecondTel = true;
+          payload.title = f2.title;
+          payload.ownerFirstname = f2.ownerFirstname;
+          payload.ownerLastname = f2.ownerLastname;
+          payload.email = f2.email;
+          payload.nickname = f2.nickname;
+
+          if (isEditing) {
+            cusCom.customerCompanyId = f2.customerCompanyId;
+            payload.customerCompany = cusCom;
+            await updateCustomerEx(payload).then((res) => {
+              if (res.success) {
+                navigate("/ShopManagementPage/CorporateShop");
+              }
+            });
+          } else {
+            payload.customerCompany = cusCom;
+            await createCustomerEx(payload).then((res) => {
+              if (res.success) {
+                navigate("/ShopManagementPage/CorporateShop");
+              }
+            });
+          }
+        })
+        .catch((errInfo) => {
+          console.log("form2 errInfo", errInfo);
+        });
     }
   };