+++ src/pages/ShopManagementPage/CorporateShop/Index.tsx
@@ -1,10 +1,5 @@
-import {
-  PlusOutlined,
-  SearchOutlined,
-  ShopOutlined,
-  UnorderedListOutlined,
-} from "@ant-design/icons";
-import { Col, Modal, Row, Table } from "antd";
+import { PlusOutlined, SearchOutlined, UnorderedListOutlined } from "@ant-design/icons";
+import { Col, Modal, Row, Table, Tag, Form } from "antd";
 import React, { useEffect, useState } from "react";
 import Button from "../../../components/Button/Button";
 import { CardContainer } from "../../../components/Card/CardContainer";
@@ -16,7 +11,8 @@ import Text from "../../../components/Text/Text";
 import { useNavigate } from "react-router-dom";
 import Input from "../../../components/Input/Input";
 import styled from "styled-components";
-import { getCusCorporate } from "../../../datasource/CustomerDatasource";
+import { checkTaxNo, getCusCorporate } from "../../../datasource/CustomerDatasource";
+import { zoneDatasource } from "../../../datasource/ZoneDatasource";
 
 const Header = styled(Col)`
   border-radius: 8px;
@@ -31,19 +27,22 @@ const Header = styled(Col)`
 function IndexCorporateShop(): JSX.Element {
   const navigate = useNavigate();
   const company = JSON.parse(localStorage.getItem("company")!);
-
+  const [form] = Form.useForm();
   const [dataState, setDataState] = useState<{ count: number; data: any[] }>({
     count: 0,
     data: [],
   });
-  const pageSize = 8;
+  const pageSize = 10;
   const [page, setPage] = useState<number>(1);
   const [search, setSearch] = useState<string>("");
-  const [isActive, setIsActive] = useState<any>("");
+  const [searchZone, setSearchZone] = useState<string>("");
+  const [cusComId, setCusComId] = useState<string>("");
 
   const [showModal, setShowModal] = useState<boolean>(false);
-  const [dataUser, setDataUser] = useState<any>({});
-  const [checkDup, setCheckDup] = useState<boolean>(true);
+  const [checkTaxId, setCheckTaxId] = useState<boolean>(false);
+  const [cusId, setCusId] = useState<string>("");
+
+  const [zone, setZone] = useState<any>([]);
 
   const getCusList = async () => {
     await getCusCorporate({
@@ -51,14 +50,23 @@ function IndexCorporateShop(): JSX.Element {
       take: pageSize,
       companyCode: `${company.companyCode}`,
       search,
+      sortField: "updateDate",
+      sortDirection: "DESC",
+      zone: searchZone,
     }).then((res) => {
       setDataState({ count: res.count_total || 0, data: res.data });
+    });
+  };
+  const getZone = async () => {
+    await zoneDatasource.getAllZoneByCompany(company?.companyCode).then((res) => {
+      setZone(res);
     });
   };
 
   useEffect(() => {
     getCusList();
-  }, [search, page]);
+    getZone();
+  }, [search, page, searchZone]);
 
   const ActionBtn = ({ onClick, icon }: any) => {
     return (
@@ -74,6 +82,7 @@ function IndexCorporateShop(): JSX.Element {
       </Col>
     );
   };
+
   const columns: any = [
     {
       title: "รหัสร้านค้า",
@@ -91,7 +100,15 @@ function IndexCorporateShop(): JSX.Element {
       key: "customerName",
       render: (value: any, row: any, index: number) => {
         return {
-          children: <Text>{value}</Text>,
+          children: (
+            <>
+              <Text>{value}</Text>
+              <br />
+              <Text level={6} color='Text3'>
+                {row?.customer?.taxNo}
+              </Text>
+            </>
+          ),
         };
       },
     },
@@ -101,7 +118,11 @@ function IndexCorporateShop(): JSX.Element {
       key: "userName",
       render: (value: any, row: any, index: number) => {
         return {
-          children: <Text>{value}</Text>,
+          children: (
+            <Text>
+              {row.customer.ownerFirstname || "-"} {row.customer.ownerLastname}
+            </Text>
+          ),
         };
       },
     },
@@ -127,8 +148,23 @@ function IndexCorporateShop(): JSX.Element {
     },
     {
       title: "สถานะ",
-      dataIndex: "status",
-      key: "status",
+      dataIndex: "isActive",
+      key: "isActive",
+      render: (value: any, row: any, index: number) => {
+        return {
+          children: (
+            <>
+              <Tag color={value ? color.success : color.error}>
+                {value ? "เปิดใช้งาน" : "ปิดใช้งาน"}
+              </Tag>
+              <br />
+              <Text level={6} color='Text3'>
+                ● {row?.zone}
+              </Text>
+            </>
+          ),
+        };
+      },
     },
     {
       title: "จัดการ",
@@ -140,7 +176,9 @@ function IndexCorporateShop(): JSX.Element {
           children: (
             <Row justify={"center"} gutter={16}>
               <ActionBtn
-                onClick={() => navigate("/ShopManagementPage/detailCorporateShop/1050")}
+                onClick={() =>
+                  navigate(`/ShopManagementPage/detailCorporateShop/${row.customerCompanyId}`)
+                }
                 icon={<UnorderedListOutlined />}
               />
             </Row>
@@ -149,6 +187,18 @@ function IndexCorporateShop(): JSX.Element {
       },
     },
   ];
+
+  const checkTax = async (e: string) => {
+    const payload = await checkTaxNo(e).then((res) => {
+      setCusId(res.responseData ? res.responseData.customerId : "0");
+      const isCreate = res?.responseData?.customerCompany?.find(
+        (c) => c.company === company.companyCode,
+      );
+      setCusComId(isCreate && isCreate.customerCompanyId);
+      return isCreate ? true : false;
+    });
+    return payload;
+  };
 
   return (
     <>
@@ -168,14 +218,44 @@ function IndexCorporateShop(): JSX.Element {
               </Col>
               <Col>
                 <Select
-                  placeholder='เขต : ทั้งหมด'
-                  data={[]}
+                  allowClear
+                  placeholder='เขตทั้งหมด'
+                  data={
+                    zone.map((z) => ({
+                      label: z.zoneName,
+                      key: z.zoneId,
+                      value: z.zoneName,
+                    })) || []
+                  }
                   style={{
                     width: 180,
                     fontFamily: "Sarabun",
                   }}
+                  onChange={(e) => {
+                    setSearchZone(e);
+                    setPage(1);
+                  }}
                 />
               </Col>
+              {/* <Col>
+                <Select
+                  allowClear
+                  placeholder='สถานะทั้งหมด'
+                  data={
+                    [
+                      //
+                    ]
+                  }
+                  style={{
+                    width: 180,
+                    fontFamily: "Sarabun",
+                  }}
+                  onChange={(e) => {
+                    setSearchZone(e);
+                    setPage(1);
+                  }}
+                />
+              </Col> */}
               <Col>
                 <Button
                   onClick={() => setShowModal(!showModal)}
@@ -195,9 +275,10 @@ function IndexCorporateShop(): JSX.Element {
           dataSource={dataState.data || []}
           columns={columns || []}
           pagination={{
+            position: ["bottomCenter"],
             current: page,
             total: dataState?.count || 0,
-            pageSize: 8,
+            pageSize: pageSize,
             onChange: (page) => {
               setPage(page);
             },
@@ -215,9 +296,8 @@ function IndexCorporateShop(): JSX.Element {
           }
           centered
           onCancel={() => {
-            //setCheckDup(true);
+            form.setFieldValue("taxNo", "");
             setShowModal(false);
-            setDataUser("");
           }}
           destroyOnClose
           cancelText={"ยกเลิก"}
@@ -225,87 +305,66 @@ function IndexCorporateShop(): JSX.Element {
           okButtonProps={{
             style: {
               color: color.white,
-              borderColor: color.primary,
-              //checkDup && Object.keys(dataUser).length > 0 ? color.primary : color.Disable,
-              backgroundColor: color.primary,
-              //checkDup && Object.keys(dataUser).length > 0 ? color.primary : color.Disable,
+              borderColor: checkTaxId ? color.primary : color.Disable,
+              backgroundColor: checkTaxId ? color.primary : color.Disable,
             },
           }}
-          onOk={() => navigate("/ShopManagementPage/createCorporateShop/create")}
+          onOk={() =>
+            cusComId
+              ? navigate(`/ShopManagementPage/createCorporateShop/${cusComId}/edit`)
+              : navigate(`/ShopManagementPage/createCorporateShop/create/${cusId}`)
+          }
           cancelButtonProps={{
             style: { color: color.primary, borderColor: color.primary },
           }}
         >
-          <Text fontWeight={600}>หมายเลขประจำตัวผู้เสียภาษี (ร้านค้า)</Text>
-          <Row>
-            <Col span={24}>
-              <Input
-                prefix={<SearchOutlined style={{ color: color.Disable }} />}
-                placeholder='ระบุหมายเลขประจำตัวผู้เสียภาษี'
-                onChange={async (e: any) => {
-                  if (!e) {
-                    setDataUser("");
-                    setCheckDup(true);
-                  } else {
-                    console.log(2);
-                  }
-                }}
-              />
-            </Col>
-          </Row>
-          <br />
-          {Object.keys(dataUser).length > 0 && checkDup ? (
-            <Header>
-              <Row gutter={16}>
-                <Col span={2}>
-                  <ShopOutlined style={{ fontSize: "20px" }} />
-                </Col>
-                <Col span={22}>
-                  <Text fontSize={16} color='primary' fontWeight={600}>
-                    {dataUser.customerName}
-                  </Text>
-                </Col>
-              </Row>
-              <Row gutter={16}>
-                <Col span={12}>
-                  <Text fontSize={14} fontWeight={500}>
-                    หมายเลขนิติบุคคล
-                  </Text>
-                  <br />
-                  <Text fontSize={14}>{dataUser.taxId}</Text>
-                </Col>
-                <Col span={12}>
-                  <Text fontSize={14} fontWeight={500}>
-                    เบอร์โทรศัพท์
-                  </Text>
-                  <br />
-                  <Text fontSize={14}>{dataUser.telephone}</Text>
-                </Col>
-              </Row>
-              <Row gutter={16}>
-                <Col span={12}>
-                  <Text fontSize={14} fontWeight={500}>
-                    เจ้าของร้าน
-                  </Text>
-                  <br />
-                  <Text fontSize={14}>{dataUser.userName}</Text>
-                </Col>
-                <Col span={12}>
-                  <Text fontSize={14} fontWeight={500}>
-                    จังหวัด
-                  </Text>
-                  <br />
-                  <Text fontSize={14}>{dataUser.province}</Text>
-                </Col>
-              </Row>
-            </Header>
-          ) : Object.keys(dataUser).length === 0 && !checkDup ? (
-            <Text color='error' fontSize={16}>
-              ร้านค้าที่เลือกมีข้อมูลแล้วในระบบ กรุณาตรวจสอบข้อมูลใหม่อีกครั้ง
-            </Text>
-          ) : (
-            <></>
-          )}
+          <Form form={form}>
+            <Text fontWeight={600}>หมายเลขประจำตัวผู้เสียภาษี (ร้านค้า)</Text>
+            <Row>
+              <Col span={24}>
+                <Form.Item
+                  name='taxNo'
+                  style={{
+                    width: "100%",
+                    marginBottom: 0,
+                  }}
+                  rules={[
+                    {
+                      required: true,
+                      message: "กรุณากรอกเลขประจำตัวผู้เสียภาษี",
+                    },
+                    {
+                      pattern: /^[0-9]{13}$/,
+                      message: "กรุณากรอกเลขประจำตัวผู้เสียภาษีให้ถูกต้อง",
+                    },
+                    {
+                      async validator(_, value) {
+                        if (value.length === 13) {
+                          const tax = await checkTax(value);
+                          setCheckTaxId(true);
+                          return tax
+                            ? Promise.reject(
+                                "หมายเลขร้านค้าที่ระบุมีข้อมูลอยู่แล้วในระบบ กรุณาตรวจสอบข้อมูลใหม่อีกครั้ง",
+                              )
+                            : Promise.resolve();
+                        } else {
+                          setCheckTaxId(false);
+                          return Promise.reject();
+                        }
+                      },
+                    },
+                  ]}
+                >
+                  <Input
+                    prefix={<SearchOutlined style={{ color: color.Disable }} />}
+                    placeholder='ระบุหมายเลขประจำตัวผู้เสียภาษี'
+                    autoComplete='off'
+                    maxLength={13}
+                  />
+                </Form.Item>
+              </Col>
+            </Row>
+          </Form>
         </Modal>
       )}
     </>